requirements.txt
absl-py==2.0.0
appnope==0.1.2
asttokens==2.0.5
attrs==23.2.0
backcall==0.2.0
blinker==1.7.0
Bottleneck==1.3.5
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
comm==0.1.2
contourpy==1.2.0
cycler==0.11.0
debugpy==1.6.7
decorator==5.1.1
exceptiongroup==1.0.4
executing==0.8.3
fastjsonschema==2.19.1
Flask<3.0,>=2.0
fonttools==4.25.0
google-api-core==2.17.1
google-api-python-client==2.121.0
google-auth==2.25.2
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.62.0
gspread==5.12.3
httplib2==0.22.0
idna==3.6
ipykernel==6.25.0
ipython==8.15.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_client==8.6.0
jupyter_core==5.5.0
kiwisolver==1.4.4
MarkupSafe==2.1.5
matplotlib==3.8.0
matplotlib-inline==0.1.6
munkres==1.1.4
nbformat==5.9.2
nest-asyncio==1.5.6
numexpr==2.8.7
numpy==1.26.2
oauthlib==3.2.2
ortools==9.6.2534
packaging==23.1
pandas==2.1.4
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.1
pip==23.3.1
platformdirs==3.10.0
plotly==5.19.0
prompt-toolkit==3.0.36
protobuf==4.25.1
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
Pygments==2.15.1
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3.post1
pyzmq==25.1.0
referencing==0.33.0
regex==2023.10.3
requests==2.31.0
requests-oauthlib==1.3.1
rpds-py==0.18.0
rsa==4.9
scipy==1.11.4
setuptools==68.2.2
six==1.16.0
stack-data==0.2.0
tenacity==8.2.2
tornado==6.3.3
traitlets==5.7.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.1.0
wcwidth==0.2.5
Werkzeug==3.0.1
wheel==0.41.2
